// bindings/typescript/types/MikanFunctionTypes.ts
// This file is auto generated. DO NOT EDIT.

export class MikanFunctionDescriptor {
  ownerSystemClass: string = '';
  ownerComponentClass: string = '';
  functionName: string = '';
  displayName: string = '';

  static __serializationMetadata: Array<{name: string, type: string, isArray?: boolean, isMap?: boolean, keyType?: string, valueType?: string}> = [
    { name: 'ownerSystemClass', type: 'string' },
    { name: 'ownerComponentClass', type: 'string' },
    { name: 'functionName', type: 'string' },
    { name: 'displayName', type: 'string' }
  ];
}